% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyapi-package.R
\docType{package}
\name{anyapi-package}
\alias{anyapi}
\alias{anyapi-package}
\title{anyapi: Quickly Access Hundreds of 'APIs'}
\description{
Tools to interactively query hundreds of web 'APIs', with documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/anyapi}
  \item \url{https://anyapi.api2r.org/}
  \item Report bugs at \url{https://github.com/jonthegeek/anyapi/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
