% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_api.R
\name{find_api}
\alias{find_api}
\title{Search for APIS}
\usage{
find_api(pattern, refresh = FALSE)
}
\arguments{
\item{pattern}{A string to search for. This string is converted to lowercase
before comparison.}

\item{refresh}{Whether to force a refresh of the APIs.guru API list.}
}
\value{
A tibble with columns \code{title}, \code{provider}, \code{service}, and \code{apid_url}.
}
\description{
Search APIs from APIs.guru by title, service, and provider.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
find_api("events")
find_api("fec.gov")
\dontshow{\}) # examplesIf}
}
